src/requirements.txt
#appdirs==1.4.4
#black==21.7b0
boto3==1.18.26
botocore==1.21.26
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
#click==8.0.1
cryptography==3.4.7
idna==3.2
jmespath==0.10.0
mypy-extensions==0.4.3
#pathspec==0.9.0
#paypal-checkout-serversdk==1.0.1
#paypalhttp==1.0.0
#paypalrestsdk==1.13.1
pycparser==2.20
pyOpenSSL==20.0.1
python-dateutil==2.8.2
regex==2021.8.21
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
stripe==2.60.0
#tomli==1.2.1
urllib3==1.26.6
